Add basic reimplementation of 'list' cmd
TODO:
- fix listing of images when an image doesn't have a name
- highlighting of running containers
- converting to human readable format creation time of images

When an image doesn't have a name, the function fails, because the
sorting functions freaks out. Listing of just containers is alright.

Highlighting is a tough one because of the way the table of images and
containers is created. We can add the shell code for the color but that
messes with the formatting. One solution is to wrap all the text with
some color and another is to use a different table formatting/rendering
solution.

Creation time conversion shouldn't be too hard to fix, but I would maybe
poke the podman guys about this, because they are slightly inconsistent
when it comes to comparing their standard output vs json. Also there are
differences in format between images and containers (cmds `images` and
`ps`)