Add logger that respects log-level flag
To have the functionality to set the type of info we want to output, I
chose to add to the project 'logrus'[1]. It's a logger library also used
by Podman that allows very good handling of output.

This required setting up the 'PersistentPreRunE' function of Cobra, that
is ran before every command (hence Persistent). In this command is
logrus set to desired level.

[1] https://github.com/Sirupsen/logrus