Stop using the 'errors' library
At first I thought we need a more powerful error handling system but now
I belive that is not really the case. So for now I'll try to handle
errors the 'go way' and see how that goes. Also the library does not
really provide that much more convenience because it basically just
wraps an error around a former error which is not really the prettiest
solution.