Upgrade function for joining JSON files
This solves the problem with sorting jsons that had empty entries. The
solution is to throw such entries out of the json while they are joined.

This solution is not ideal, because it only checks entries called
'names' and 'Names'. Ideally it would check any entry and it could be
handled by a parameter.