Add the base of 'create' command
This is not yet a functional command. A lot of implementation details
are missing.

Currently what IS done:
- identification of host system and it's version (tested only on Fedora)
- checking if an image is on the machine
- preparation of dbus system bus address
- checking if--ulimit host' is supported

This proposes a change in the design of the command. Instead of
specifying the name of the container with a flag ('-c | --container')
the name of the container is now specified as an argument.