Add handling of default naming
This moves the logic for resolving default naming of containers and
images to a separate function. By default the used image should reflect
the host. If the host is unknown then 'fedora-toolbox' will be chosen.

The default name of the container should reflect the name of the used
image. If the image name has a colon then it is replaced by a dash.

This also adds a function that checks if a container name matches the
required regexp pattern.