Handle errors in 'rm' in a different way
Commands in Cobra should return error and based on that Cobra can handle
the return value.

When 'rm' can't find a container then it says 'Failed to inspect
container' which is rather confusing to a person who is not used to
Podman. According to Podman manpage the value 125 is returned ' if the
error is with podman itself'. So I called it 'Internal Podman error' and
in the parent function I catch it if errors.As().

This will probably lead to creatino of a separate package for handling
the types of Errors to prevent creating the errors manually (line 80).