Finish (almost) 'create' command
With this the 'create' command should create fully working containers.
The only piece missing is the part for checking/pulling images if an
image is not on the host.

The entry point of contaners is still taken from the former
implementation. So if you want to use this version you need to have the
old toolbox in your PATH before this version.