Partially implement 'run' command
The 'run' command should in general work with this commit. There are
still few parts missing eg. checking if a command exists in a container
(useful mainly for the 'enter' command), offering to create a container
if it does not exist.

Also the parsing of command line options doesn't allow passing --flags
because they are interpreted by toolbox flags. This needs to be fixed.