Fix functions for interaction with mounts
I made a mistake and to get the mount point of a target I used a tool
called findmnt even though the output is not correct. So instead I went
back to df. But still there should be a native way of getting this info.

Also sometimes it happens that the output is more than just a line which
I didn't expect when I wrote the functions. This should fix such cases.