Reimplement the 'init-container' command
This should be a fully working entry-point for containers.

I removed the part where we link '/run/host/etc/***' to '/etc/***'
because that should be done in the later part thanks to the call to
flatpak-session-helper.

I also included a fix for the case when the path /run/user/1000/toolbox
does not exist. This problem was not noticeable in the old version
because there is migrate() function (that we no longer need) that took
care of it without us properly realizing.

For now the mountBind function binds by execing the 'mount' utility but
in future versions we want to use the syscall.Mount function.