Make 'create' aware of the toolbox implementation
Because go prefers camelCase name convention and the Cobra library
follows that, I renamed the 'init-container' command to 'initContainer'
(could be also renamed just to 'init'). Also the new implementation does
not accept the '--verbose' flag but uses 'log-level'. This causes an
incompatibility between the two implementations. To solve this I added
this to resolve what implementation is the system default (according to
PATH), set the arguments and flags appropriately and then bind the
appropriate toolbox to /usr/bin/toolbox in the container.

Thanks to that you can affect what implementation is used by 'create' by
injecting to PATH.

If no toolbox executable is found in PATH then the executable itself is
used.