Add the 'verbose' flag option
The new implementation uses 'log-level' instead of 'verbose' because the
flexiblity is very nice. But containers created with the shell toolbox
require the 'verbose' flag.

If 'verbose' is used, 'log-level' is set to 'debug'. And the option is
marked as deprecated -> a message about the deprecation is showed when
the option is used.