Restrict usage of Toolbox in containers
Currently there is no support for 'flatpak-spawn --host' invocation
which is crucial for running Toolbox in containers (because even when
invoked in a container the command is actually ran on the host system).

This restriction only doesn't apply to the 'init-container' for obvious
reasons.

In the future this restriction will be lifted.