Add handling for Podman migration
When Podman updates it is sometimes necessary to call the `podman system
migrate` to ensure that the existing containers will run properly with
the new version.

Toolbox checks the need to update by storing the latest version of
Podman used by Toolbox in ~/.config/toolbox/podman-system-migrate file.

The function doesn't have to be invoked when the 'init-container'
command is used or if Toolbox is ran in a container (because the
container should already be alright, right?).