Make pulling of images cleaner and more reliable
With this the output of `podman pull` is hidden and replaced by a
spinner (that should not destroy your terminal when pulling is aborted
in the middle).

We still only support our own Toolbox images (we want to change this in
the future) so until anything changes the only trusted registry is
'registry.fedoraproject.org'.

Sometimes the fedoraproject registry returns 503 HTTP error which
usually does not imply that the service is down (well.. it should but it
happens quite frequently nowaydays). So there is a loop that invokes
`podman pull` until it at least gets a different error than the 503.