Update system tests
New workflow is required to set up toolbox so the CI has to be adjusted.

This drops testing on Fedora 30 (for now) because there is an older
version Golang that does not support the package 'errors' that came in
Go 1.13.

Also the UI of Toolbox has slightly changed so the tests needed to be
adjusted.