Drop unnecessary compatibility check
I thought it would be a good idea to check which version of Toolbox is
the system default but it actually is not. The previous implementation
always considers the called binary (resp. script) to be the default
disregarding what is first in PATH.

So if you have the 'toolbox' binary in /bin or /usr/bin but call it from
~/go/bin then TOOLBOX_PATH will be /home/user/go/bin/toolbox (or
something else depending on your system).