Add 'default | d' options to the 'run' cmd
With the rewrite I took the liberty to slightly change the way the 'run'
command is used. It is now more similar to the way Podman handles its
'exec' command. But with the change the convenience to just type
'toolbox run <your-cmd>' and expect it to run in the default container
has been lost. To make up for it at least a bit I added the 'default |
d' option. The usage of 'run' cmd looks now like this:

- `toolbox run echo 123` means: run cmd '123' in container echo

- `toolbox run -d echo 123` means: run cmd 'echo 123' in the default
  container

- `toolbox run -r 31 echo 123` means: run cmd 'echo 123' in container
  with release 31

Also with this are added fatal messages if bad number of arguments is
provided.