playbooks: Drop task for pulling images before tests
When tests were ran then quite often they would fail because of problems
when pulling images from the fedora registry. This was compensated for
with a pre-job ran in the CI before the tests were executed.

But that scenario is not ideal. The rewriten Toolbox tries several times
to pull an image in case the transfer fails before it throws an error.
Thanks to that the pre-job is no longer needed.

But still huge thanks to Tristan who introduced the patch!

This reverts commit: 50683c9d9a78adc9378e7746915f435e59e6e970