Update bash completion
This is based on #304. The completion should work ok for basic usage but
it can still be quite easily broken. Please don't do that :)!

In the future I would like to use autogenerated bash completion made by
the Cobra framework but before that happens we'll keep on using this
script.

Another things that I want to change are the functions for getting
containers and images. Right now it returns ALL images and ALL
containers even though they cannot be used with Toolbox. So in the
future I'll try to implement some kind of --field switch into the `list`
command .