Add generator of manual pages
With the move to Go we no longer need to use a build system because the
Go ecosystem has all the tools necessarry for building the tool. Also
generating manual pages, bash-completion and many other resources can be
done by the tools themselves (even though it feels even to me as
reinventing the wheel a bit).

So with this Toolbox should be capable of generating it's own manpages
out of the provided markdown files. Flags --input-folder and
--output-folder must be set.