Drop Meson
Meson is no longer needed to build the project because Go ecosystem is
selfsustainable. Of course there is a tradeof: no simple local
installation. If it proves to be too big of a problem then the change
can be reverted and the build process adjusted.

Installation of all support files is now handled by the .spec file in
the Toolbox repo in Fedora distgit[0].

[0] https://src.fedoraproject.org/rpms/toolbox