Make pulling images smarter and move code around
I realized a problem with the '--image' flag in 'create' cmd that it
prepends 'registry.fedoraproject.org' to the name of an image regardless
if there already a domain is. Toolbox by default trust only the Fedora
registry but if a user wants to use own image then Toolbox should allow
it. So with this if an image contains a full path (has a domain) then
Toolbox will accept it.

At the same time I split some parts of the code into separate functions
because the main function is already quite big.

I capped the number of retries when 'podman pull' fails with the 503
http code to 3.

Until now when the '--log-podman' option was used the spinner "Pulling
image X" was hidden. This is not a good experience because the function
for pulling images shows the output of 'podman pull' *after* it is done
because of proper error handling (if that's a wrong approach then please
fix it for me). So from now on the spinner will be always shown.