Make most errors in 'init-container' non-fatal
Until now almost any problem during initialization of a container would
cause a fatal error. Because of that the container would not be
enterable.

Such a scenario is not very nice because such container should be stil
at least partially usable.

Still, if there is a problem then Toolbox itself is not capable, yet, of
showing the problem (and maybe it's possible solution). To investigate
you still need to use `podman logs`.