Use fixed module for comparing version numbers
The module that we used until now[0] had a regression where a function
returned a different value than described in the description. Because
the repository is archived I forked the repo[1], fixed the issue, made a
new release and converted the code to use the forked module.

So now the CheckVersion function in package 'podman' returns 1 if the
required version is higher than Podman's and -1 if it the other way
around (so it behaves more like C's strcmp).

[0] github.com/mcuadros/go-version
[1] github.com/HarryMichal/go-version