cmd/initContainer: Follow symlinks when redirecting
There has been a long-standing issue[0] when symlinking /etc/resolv.conf
to /run/host/etc/resolv.conf (host's resolv.conf). Many solutions were
proposed. Hardcode other paths, where the file could be, completely rely
on tools like flatpak-session-helper that will track such files in a
single directory, or update the symlinking function to follow symlinks
and update the target path.

The first solution is the easiest short-term but not very good
long-term. The second solution is possibly the best long-term but there
is a problem with using flatpak-session-helper. It cannot be used as
root. That leaves the third option, follow symlinks and updating the
target path.

This commit takes the third approach to solve the issue. Now the target
of the symlinking is first tested if it is a symlink and if it is, it's
target will be set as the new target. This is repeated if the new target
is also a symlink. The function looks for the links in the chain under
/run/host where the host's filesystem is available. Relative symlinks
are handled by this (thanks Tudor Roman!).

Based on: https://github.com/containers/toolbox/pull/380

[0] https://github.com/containers/toolbox/issues/187