cmd/initContainer, pkg/utils: Add symlink healthcheck
Previous commit adds support for traversing symlink targets in the
redirectPath function. It has one shortcoming, though. It does not fix
already broken links (e.g., in the case of upgrading from previous
releases to Fedora 33 that enables systemd-resolved that substitutes
/etc/resolv.conf with a symlink).

This commit adds a function FollowSymlink to the utils package that
follows a possible chain of symlinks (symlink leading to another
symlink). It accepts an argument that makes it search for the target in
the host's filesystem available in a toolbox container under /run/host.
It does that job by prepending '/run/host' to the resolved symlink
target. The function has a failsafe mechanism for detecting cyclic
symlinks.

The healthcheck mechiansm is added to the redirectPath function, making
redundant symlink checks before calling the function.

Current behaviour of the redirectPath with the healthcheck mechanism:

  - Check the validity of the target - no point in symlinking to an
    invalid location - check is done by traversing the symlink chain
  - Continue working with last piece of the symlink traversal - it's
    impossible to simply create a symlink to the first piece of the
    chain because the symlink chain is not adjusted to point to the
    host's filesystem in the container under /run/host
  - Check the validity of the container symlink and compare it's target
    with the last piece of the target symlink traversal
  - Container symlink is valid and matches the requested target -> skip
    redirection
  - Proceed with the redirection (creating the symlink) - delete the
    container path and create a new symlink